requirements.txt
torch==2.0.1
torchvision==0.15.2
transformers==4.30.2
datasets==2.13.1
pillow==9.5.0
opencv-python==4.8.0.74
matplotlib==3.7.2
numpy==1.25.1
scikit-learn==1.3.0
tqdm==4.65.0
pyyaml==6.0.1
albumentations==1.3.1
wandb==0.15.5 